# Food-Delivery/api_gateway_service/docs/swagger.yaml
basePath: /
definitions:
  courier.Courier:
    properties:
      courier_id:
        description: Kuryerning noyob ID'si
        type: string
      email:
        description: Kuryerning elektron pochtasi
        type: string
      name:
        description: Kuryerning ismi
        type: string
      phone_number:
        description: Kuryerning telefon raqami
        type: string
      status:
        allOf:
        - $ref: '#/definitions/courier.CourierStatus'
        description: Kuryerning holati
    type: object
...
  courier.CourierResponse:
    properties:
      courier:
        allOf:
        - $ref: '#/definitions/courier.Courier'
        description: Kuryer ma'lumotlari (Create va Get operatsiyalari uchun)
      message:
        description: Javob xabari
        type: string
      success:
        description: Amaliyot muvaffaqiyatli bo'lganligini bildiruvchi flag
        type: boolean
    type: object
  courier.CourierStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      CourierStatus_AVAILABLE: 'Kuryer holati: mavjud'
      CourierStatus_ON_DELIVERY: 'Kuryer holati: yetkazish jarayonida'
      CourierStatus_UNAVAILABLE: 'Kuryer holati: mavjud emas'
    x-enum-varnames:
    - CourierStatus_AVAILABLE
    - CourierStatus_UNAVAILABLE
    - CourierStatus_ON_DELIVERY
...
  courier.CreateCourierRequest:
    properties:
      email:
        description: Kuryerning elektron pochtasi
        type: string
      name:
        description: Kuryerning ismi
        type: string
      phone_number:
        description: Kuryerning telefon raqami
        type: string
      status:
        allOf:
        - $ref: '#/definitions/courier.CourierStatus'
        description: Kuryerning holati
    type: object
...
  courier.UpdateCourierRequest:
    properties:
      courier_id:
        description: Kuryerning noyob ID'si
        type: string
      email:
        description: Kuryerning elektron pochtasi
        type: string
      name:
        description: Kuryerning ismi
        type: string
      phone_number:
        description: Kuryerning telefon raqami
        type: string
      status:
        allOf:
        - $ref: '#/definitions/courier.CourierStatus'
        description: Kuryerning holati
    type: object
...
  notification.AdminAlert:
    properties:
      admin_id:
        description: Ogohlantirishni oladigan admin ID'si
        type: string
      alert_id:
        description: Ogohlantirish ID'si
        type: string
      created_at:
        description: Ogohlantirish yaratilgan vaqt
        type: string
      is_read:
        description: Admin ogohlantirishni o'qiganligini bildiradi
        type: boolean
      message:
        description: Ogohlantirish matni
        type: string
      type:
        allOf:
        - $ref: '#/definitions/notification.AdminAlertType'
        description: Ogohlantirish turi
    type: object
  notification.AdminAlertResponse:
    properties:
      alert:
        allOf:
        - $ref: '#/definitions/notification.AdminAlert'
        description: Admin ogohlantirishi ma'lumotlari (Create operatsiyasi uchun)
      message:
        description: Javob xabari
        type: string
      success:
        description: Amaliyot muvaffaqiyatli bo'lganligini bildiruvchi flag
        type: boolean
    type: object
  notification.AdminAlertType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      AdminAlertType_COURIER_ACTIVITY: Kuryer faoliyati ogohlantirishi
      AdminAlertType_NEW_ORDER: Yangi buyurtma ogohlantirishi
      AdminAlertType_STATUS_CHANGE: Holat o'zgarishi ogohlantirishi
    x-enum-varnames:
    - AdminAlertType_NEW_ORDER
    - AdminAlertType_STATUS_CHANGE
    - AdminAlertType_COURIER_ACTIVITY
...
  notification.CreateAdminAlertRequest:
    properties:
      admin_id:
        description: Ogohlantirishni oladigan admin ID'si
        type: string
      message:
        description: Ogohlantirish matni
        type: string
      type:
        allOf:
        - $ref: '#/definitions/notification.AdminAlertType'
        description: Ogohlantirish turi
    type: object
...
  notification.UpdateAdminAlertRequest:
    properties:
      alert_id:
        description: Ogohlantirish ID'si
        type: string
      is_read:
        description: Yangilangan o'qilganligi holati
        type: boolean
      message:
        description: Yangilangan ogohlantirish matni
        type: string
      type:
        allOf:
        - $ref: '#/definitions/notification.AdminAlertType'
        description: Ogohlantirish turi (yangilanishi mumkin)
    type: object
...
  product.Cart:
    properties:
      cart_id:
        description: Savatning noyob ID'si
        type: string
      created_at:
        description: Savat yaratilgan vaqti
        type: string
      user_id:
        description: Savatga ega foydalanuvchi ID'si
        type: string
    type: object
...
  product.CartListResponse:
    properties:
      carts:
        description: Savatlar ro'yxati
        items:
          $ref: '#/definitions/product.Cart'
        type: array
    type: object
  product.CartResponse:
    properties:
      cart:
        allOf:
        - $ref: '#/definitions/product.Cart'
        description: Savat ma'lumotlari (Create va Get operatsiyalari uchun)
      message:
        description: Javob xabari
        type: string
      success:
        description: Amaliyot muvaffaqiyatli bo'lganligini bildiruvchi flag
        type: boolean
    type: object
...
  product.CreateCartRequest:
    properties:
      user_id:
        description: Savatga ega foydalanuvchi ID'si
        type: string
    type: object
...
  product.UpdateCartRequest:
    properties:
      cart_id:
        description: Savatning noyob ID'si
        type: string
      user_id:
        description: Savatga ega foydalanuvchi ID'si
        type: string
    type: object
...
info:
  contact: {}
  description: API for Food Delivery System resources
  title: Food Delivery System API
  version: "1.0"
paths:
  /api/alerts/admin:
    post:
      consumes:
      - application/json
      description: Add a new admin alert to the system
      parameters:
      - description: Alert Data
        in: body
        name: alert
        required: true
        schema:
          $ref: '#/definitions/notification.CreateAdminAlertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/notification.AdminAlertResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new admin alert
      tags:
      - Alerts
  /api/alerts/admin/{alertId}:
    delete:
      description: Remove an admin alert from the system by its ID
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.AdminAlertResponse'
        "400":
          description: Invalid Alert ID
          schema:
            type: string
        "404":
          description: Alert Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete admin alert by ID
      tags:
      - Alerts
    get:
      description: Retrieve details of an admin alert by its ID
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.AdminAlertResponse'
        "400":
          description: Invalid Alert ID
          schema:
            type: string
        "404":
          description: Alert Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get admin alert by ID
      tags:
      - Alerts
    put:
      consumes:
      - application/json
      description: Update details of an existing admin alert by its ID
      parameters:
      - description: Alert ID
        in: path
        name: alertId
        required: true
        type: string
      - description: Alert Data
        in: body
        name: alert
        required: true
        schema:
          $ref: '#/definitions/notification.UpdateAdminAlertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.AdminAlertResponse'
        "400":
          description: Invalid Alert ID
          schema:
            type: string
        "404":
          description: Alert Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update admin alert by ID
      tags:
      - Alerts
...
  /api/carts:
    get:
      description: Retrieve a list of all carts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.CartListResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all carts
      tags:
      - Carts
  /api/carts/{cartId}:
    delete:
      description: Remove a cart from the system by its ID
      parameters:
      - description: Cart ID
        in: path
        name: cartId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cart deleted successfully
          schema:
            type: string
        "400":
          description: Invalid Cart ID
          schema:
            type: string
        "404":
          description: Cart Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete cart by ID
      tags:
      - Carts
    get:
      description: Retrieve details of a cart by its ID
      parameters:
      - description: Cart ID
        in: path
        name: cartId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.CartResponse'
        "400":
          description: Invalid Cart ID
          schema:
            type: string
        "404":
          description: Cart Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get cart by ID
      tags:
      - Carts
    put:
      consumes:
      - application/json
      description: Update details of an existing cart by its ID
      parameters:
      - description: Cart ID
        in: path
        name: cartId
        required: true
        type: string
      - description: Cart Data
        in: body
        name: cart
        required: true
        schema:
          $ref: '#/definitions/product.UpdateCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.CartResponse'
        "400":
          description: Invalid Cart ID
          schema:
            type: string
        "404":
          description: Cart Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update cart by ID
      tags:
      - Carts
  /api/carts/create:
    post:
      consumes:
      - application/json
      description: Add a new cart to the system
      parameters:
      - description: Cart Data
        in: body
        name: cart
        required: true
        schema:
          $ref: '#/definitions/product.CreateCartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/product.CartResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new cart
      tags:
      - Carts
...
  /api/couriers/{courierId}:
    delete:
      description: Remove a courier from the system by their ID
      parameters:
      - description: Courier ID
        in: path
        name: courierId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Courier deleted successfully
          schema:
            type: string
        "400":
          description: Invalid Courier ID
          schema:
            type: string
        "404":
          description: Courier Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete courier by ID
      tags:
      - Couriers
    get:
      description: Retrieve details of a courier by their ID
      parameters:
      - description: Courier ID
        in: path
        name: courierId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.CourierResponse'
        "400":
          description: Invalid Courier ID
          schema:
            type: string
        "404":
          description: Courier Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get courier by ID
      tags:
      - Couriers
    put:
      consumes:
      - application/json
      description: Update details of an existing courier by their ID
      parameters:
      - description: Courier ID
        in: path
        name: courierId
        required: true
        type: string
      - description: Courier Data
        in: body
        name: courier
        required: true
        schema:
          $ref: '#/definitions/courier.UpdateCourierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.CourierResponse'
        "400":
          description: Invalid Courier ID
          schema:
            type: string
        "404":
          description: Courier Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update courier by ID
      tags:
      - Couriers
  /api/couriers/create:
    post:
      consumes:
      - application/json
      description: Add a new courier to the system
      parameters:
      - description: Courier Data
        in: body
        name: courier
        required: true
        schema:
          $ref: '#/definitions/courier.CreateCourierRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/courier.CourierResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new courier
      tags:
      - Couriers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"